move seed data out of main crate
